% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering_method.R
\name{runCNMF}
\alias{runCNMF}
\title{Title}
\usage{
runCNMF(
  mRNAexp,
  miRNAexp,
  lncRNAexp,
  methylation,
  CNVmatrix,
  clusterNum = 3,
  nrun = 30
)
}
\arguments{
\item{mRNAexp}{A matrix of mRNA expression}

\item{miRNAexp}{A matrix of miRNA expression}

\item{lncRNAexp}{A matrix of lncRNA expression}

\item{methylation}{A matrix of methylation}

\item{CNVmatrix}{A matrix of Copy number variation}

\item{clusterNum}{Number of cluster}

\item{nrun}{The number of times the nmf algorithm has been run to compute an object}
}
\value{
return a ....
}
\description{
Title
}
\examples{
data(mRNAexp)
data(miRNAexp)
data(miRNAexp)
data(miRNAexp)
data(miRNAexp)
result=runCNMF(mRNAexp,miRNAexp,miRNAexp,miRNAexp,miRNAexp,clusterNum=3,clusterNum=1,nrun=30)
result$group
}
